+++ decompile/General/LOAD/LOAD_16_DramFileCallback.c
@@ -19,7 +19,7 @@ void DECOMP_LOAD_DramFileCallback(struct LoadQueueSlot* lqs)
 	{	
 		struct DramPointerMap* dpm = (struct DramPointerMap*)&realFileBuf[ptrMapOffset];
 	
-		DECOMP_LOAD_RunPtrMap((int)realFileBuf, (int*)DRAM_GETOFFSETS(dpm), dpm->numBytes>>2);
+		DECOMP_LOAD_RunPtrMap(realFileBuf, (int*)DRAM_GETOFFSETS(dpm), dpm->numBytes>>2);
 		
 		DRAM_SET_PATCHED(fileBuf);
